internal/doc/python/requirements.txt
python-docx
fastapi
uvicorn
pillow